robust_taxi/requirements.txt
Flask==3.0.0
pymongo==4.6.0
flask-socketio==5.3.5
python-socketio==5.10.0
flask-cors==4.0.0
werkzeug==3.1.3
eventlet
gunicorn
